chore!: refactor all serializers into their own modules
