+++ .pre-commit-config.yaml
@@ -3,4 +3,6 @@ repos:
     rev: v0.1.3
     hooks:
     - id: ruff
+      args: [ --fix ]
       exclude: (migrations)
+    - id: ruff-format
