+++ api/serializers/finance.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from finance.models import Payment
+
+
+class PaymentSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Payment
+        fields = "__all__"
